Usando funções nos exercícios

+++ modulo1/semana02/exercicio06/index.js
@@ -1,8 +1,8 @@
 const operador = prompt(
   'Digite com qual operação deseja realizar (+) (-) (*) (/)'
 )
-const numero1 = parseInt(prompt('Digite o primeiro número'))
-const numero2 = parseInt(prompt('Digite o segundo número'))
+const numero1 = parseFloat(prompt('Digite o primeiro número'))
+const numero2 = parseFloat(prompt('Digite o segundo número'))
 
 if (operador === '+') {
   alert(`O resultado do seu cálculo é: ${numero1 + numero2}`)
